Fix PPC64LE system linkage for stack arguments
Previously, there was a hack in place that would cause the caller's
parameter save area to not be used for parameters on PPC64LE. This was
due to the 64-bit ELF V2 ABI Specification making the presence of the
parameter save area optional if all parameters can be passed in
registers. However, this inadvertently broke the ability to load
parameters that were passed on the stack, since they would not map
correctly to the caller's parameter save area.

This has now been corrected by ensuring that the parameter save area is
always used if one or more parameters were passed on the stack.

Signed-off-by: Ben Thomas <ben@benthomas.ca>