Fix missing volatile in CS large object copy
When retrying to win a chunk to copy, to be sure we properly re-read the
the fields of the control word, the address of control word must be of
type pointer to volatile word.

Without it we may (for example) end up missing just incremented
outstandingCopies count and eventually hit negative value on way down
when decrementing.

The missing volatile is regression from relatively recent re-org for
CR/nonCR runtime feature: https://github.com/eclipse/omr/pull/4627
Original code indeed used to read the control word through
volatile MutableHeaderFields::slot *

Fixes: https://github.com/eclipse/openj9/issues/8478

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>