Use VFT test for hierarchy guard converted from profiled guard
When a method is inlined with profiled guard with VFT test, optimizer
may assume the object has a fixed type in the inlined body and thus
optimizing the code based on that assumption. Under some circumstances,
VP will transform a profiled guard with VFT test to hierarachy guard
with method test. Method test is too specific and insufficient to guard
other optimization that may have been done based on the assumption of
the object type.

Fix this by using VFT test.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>