Fix DeadTreeElimination for volatile load
We want a volatile field (both static and instance) access to be a
barrier to code motion of other field accesses (to preserve the ordering
semantics that are expected from volatile keyword). The way this is
enforced in the compiler is to consider volatile accesses as "defs"
essentially and relying on the aliasing on the symbol reference to kill
other symbol references impacted by the ordering semantics. Defs in the
compiler must be either a treetop or the first child under a treetop for
optimizations to locate them, i.e. there are optimizations that walk the
IL trees only looking at the treetop and first child for potential defs.
So if a volatile access was accessed for the first time at an arbitrary
spot (level) in a tree (rather than at the treetop level) then it might
get missed as a def and this may lead to incorrect optimization. This
commit ensures that volatile sym refs are found where the compiler
expects defs, namely at the appropriate treetop level.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>