Implement bmul and smul evaluators on Z
After enabling all the sub-int32 tests we now expose new problems
with the missing implemetnations of the bmul and smul evaluators. The
tests for these evaluators were merged while the sub-int32 tests on Z
were disabled and as such the test failures were never encountered.

We fix this problem by providing a trivial implementation of these
evaluators on Z.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>