Fix ILGen GPF if last block is malformed
When generating code, if the last block being generated does not end in
a control-flow statement a GPF can occur in the CFG trying to add an
appropriate exit edge. While such an unterminated block is ilformed, the
block may be unreachable and will be cleaned up by the optimizer. This
change is to prevent the compiler crashing - an edge from the last block
is added to the end of the CFG.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>