Support Known Object Table at JITServer
Add KnownObjectTable::updateKnownObjectTableAtServer()
which is used to update KOT at the server.
Remove getfPointerLocationAt() which is not used any where.

Add Compilation::isOutOfProcessCompilation()
which returns true only when it’s called at the server.
Compilation::isOutOfProcessCompilation() is used to prevent KOT
from being used in OMR code at the server
where the object pointer needs to be dereferened.

Replace the direct dereference of getPointerLocation() with getPointer().
When getPointer() is called at the server,
it sends a message to the client to dereference it
with the proper VM access.

Although there are a few places the known object pointers
are dereferenced directly, it shouldn't cause any issue
if the critical section is created
with VMAccessCriticalSection::tryToAcquireVMAccess,
because at the server the VM access will not be granted.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>