RISC-V: implement 'far' branches
On RISC-V, range of conditional branches (B-type instruction) is only
+- 4KiB. Range of unconditional branch (jal) is +- 1MiB. In case jump
target of an conditional branch is outside its reach, the conditional
branch is transformed into into a flipped conditional branch around an
unconditional ('far') branch.

This commit is a shameless copy of same mechanism introduced for POWER
in commit 0fbf852b.

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>