Enhance JitBuilder Call sample to be more exhaustive
In this change we rewrite the Call sample with a few enhancements to
introduce all call types:

- JIT to JIT
- JIT to native
- JIT to computed native
- Native to JIT

The test computation is still the same, i.e. we still compute a double
sum, however we do it in several different ways so as to test calls
both to and from JIT/native to ensure all paths are working as expected.

This is particularly useful for platforms which have special handling
of function pointers, such as AIX and z/OS.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>