Fix loading of double arguments on Z
The path for loading double arguments for outgoing calls was disabled
within the `J9_PROJECT_SPECIFIC` ifdef, and as such double arguments
were not being loaded for outgoing calls.

We fix this by moving the case label outside of the ifdef. In addition
there is no need to check for `getNumIntegerArgumentRegisters` before
incrementing the GPR index, similarly to the integer cases since the
`getIntegerArgumentRegister` API will already return the default value
if the register index does not participate in the calling convention.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>