Create the XPLINK Function Descriptor for OMR compiled bodies on z/OS
System linkage OMR compiled function bodies need to come with a proper
function descriptor to adhere to XPLINK linkage on z/OS. Currently OMR
compiled functions return to the user the entry point of the function
(function's address), however on z/OS we should instead be returning
the Function Descriptor data structure.

This change implements the Function Descriptor as a snippet in the
compiled body, similar to the PPA1 and PPA2 snippets. The descriptor is
then baked into the function literal pool, and when we emit the snippet
we will update the method address which is used by external APIs.

When dispatching to other native methods (JIT or C-native) we must also
call through the function descriptor.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>