Fix bug in value propagation for lowestOneBit
The functions `constrainIntegerLowestOneBit` and `constrainLongLowestOneBit`
were passing the functions `integerToPowerOf2` and `longToPowerOf2`,
respectively, to the helper function
`constrainHighestOneBitAndLeadingZerosHelper`. This caused the helper to compute
the highest one bit instead of the lowest one bit.  This has been fixed by
creating new functions which compute the lowest one bit for integer and log
types.

- Created new helper functions to compute lowest one bit for integer and long
  types.
- Pass these new helper functions to
  `constrainLowestOneBitAndTrailingZerosHelper`.

Fixes: https://github.com/eclipse/openj9/issues/7088

Signed-off-by: Ryan Shukla <ryans@ibm.com>