Fix localCSE where it was commoning finals in VOLATILE_ONLY Pass
For the optimization level hot and above, we run two passes in localCSE.
First one for the volatiles only and second one for non volatiles only.
In case of the volatiles only pass, it was commoning the node that are final fields/
gloabal Autos/Parms as well with the assumption that we can rely on that node to not get changed.
Issue occurs, where it encounters first reference of such non volatile nodes and just adds it to
the list of replaced nodes without replacing it. This causes an issue with the
subsequent reference of such non volatiles which will be replaced with the available
expression as node is already in the list of replaced nodes.
This commit fixes the issue so that it puts node to the replaced list only if commoning is done.

Fixes: https://github.com/eclipse/openj9/issues/8637

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>