Support Known Object Table in JITServer
This is the first step of the clean-up:
- Add `getOrCreateIndex()` and `getOrCreateIndexAt`() which
will be used to replace `getIndex()` and `getIndexAt()` in
the next clean-up phase. The reason of this change required
is to make it clear to the callers of these functions
that these functions change the known object table size.
They are not simple getters.
- Remove `virtual` declaration of the methods
since `KnownObjectTable` is an extensible class.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>