Fix reference to global align
This code is intending to refer to ::align() (i.e. a function in the global
namespace, defined in compiler/runtime/Alignment.cpp).

However since the code is itself in the OMR namespace and the align()
is unqalified the compiler will choose OMR::align() in compilation units
that include that definition, and that definition is incompatible both
lexically and functionally.

Unfortunately ::align() and Alignment.cpp are referenced in downstream
projects and so we can't just fix it in the most sensible way without
breaking things first, so instead just make sure this use of align is
properly qualified.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>