Add Azure Pipelines build definitions for Windows, Linux, and macOS
This change adds the Azure Pipelines YAML definition which builds and
tests OMR on x86 Windows, Linux, and macOS. The pipeline uses the CMake
build system exclusively and runs the same level of testing as the
Jenkins builds for the respective configurations. The Azure Pipelines
tests run on every merge to the master repository as well as all pull
requests.

1. Example of a build when a change is merged:
   https://github.com/mw918/omr/commits/master

2. Example of a build when a pull request is opened or force pushed to:
   https://github.com/mw918/omr/pull/8

In addition to running all the testing the pipeline definitions being
contributed as part of this pull request support caching via Azure
Pipelines Cache jobs. For the purposes of OMR builds `ccache` is used.
Caching is enabled by default for Linux and macOS. With caching enabled
we are able to build and test (~200k unit tests) in under 5 minutes on
those platforms. The following table describes the build and test times
when caching is enabled vs. disabled:

| Platform | Azure Caching Enabled | Azure Caching Disabled | AppVeyor | Travis CI |
|----------|-----------------------|------------------------|----------|-----------|
| Windows  | 17m 13s               | 17m 50s                | 20m 46s  |           |
| Linux    | 4m 35s                | 12m 19s                |          | 11m 15s   |
| macOS    | 5m 7s                 | 14m 44s                |          |           |

Examples of the two builds (caching enabled vs. disabled) can be found
in [1] and [2] respectively. As we can see with caching enabled Azure is
able to build and test OMR almost 3x faster than other CIs all while
having a much larger test coverage. If we compare apples to apples, i.e.
Travis CI build only vs. Azure build only then Azure is almost 10x
faster at building OMR with caching enabled.

Azure Pipelines caches are immutable by design, meaning that the
`ccache` will not be refreshed on every build with new changes. This is
a problem as our codebase will encounter more and more `ccache` misses
as changes are contributed. To alleviate this problem an
`azure-pipelines.cache` file is also checked in as part of this PR.
This file contains the SHA of the commit for which the `ccache` was
populated with. The cache for Linux and macOS is separate. This gives
us the ability to invalidate the `ccache` cache by simply changing this
file to contain the SHA of the latest commit. This can be done on an
as-per-needed basis or can be automated (for example every _n_ number
of days).

Contribution and enablement of Azure Pipelines builds to OMR has
several benefits:

- Azure has a much larger capacity (10 parallel builds per repository)
than AppVeyor and Travis CI
- Build times are much faster on Azure, meaning developers get faster
feedback on their changes
- Our current pipeline tests the same set of tests as our Jenkins
builds, where as neither AppVeyor nor Travis CI perform any tests

Overall, using Azure as another build and test cloud gives further
testing and faster feedback for OMR contributors at no cost to the
project. We believe enabling Azure Pipeline builds will be an overall
benefit to the community as a whole.

[1] https://dev.azure.com/eclipse-165/omr-windows/_build/results?buildId=434&view=results
[2] https://dev.azure.com/eclipse-165/omr-windows/_build/results?buildId=436&view=results

Closes: #4777

Co-authored-by: Filip Jeremic <fjeremic@ca.ibm.com>

Signed-off-by: Michael Wang <mw918@hotmail.com>