Move MakeVector into CodeGenTest.hpp
In the compiler unit tests, the MakeVector class was being used to
create vectors using variadic macros to avoid having to use initializer
lists (which not all of our build compilers support). However, this was
previously defined in BinaryEncoder.cpp. Since this is more general, it
has now been moved into the CodeGenTest.hpp header file.

Signed-off-by: Ben Thomas <ben@benthomas.ca>