Fix corner cases in TR::Instruction::remove
Previously, the TR::Instruction::remove method did not correctly handle
removing the first or last instruction, as it always assumed that the
return values of getNext() and getPrev() would be non-null. This has now
been corrected, and it should now be possible to remove the first or
last instruction from the instruction stream using this method.

Signed-off-by: Ben Thomas <ben@benthomas.ca>