Fix fixedSeqMemAccess using MemSrc1 for addi
Previously, the fixedSeqMemAccess helper would incorrectly generate addi
instructions using a PPCMemSrc1Instruction since addi is neither a load
nor a store. This was conceptually incorrect, as the register of an addi
instruction in this context is a target register. This has been
corrected so that PPCTrg1MemInstruction is used instead.

Signed-off-by: Ben Thomas <ben@benthomas.ca>