Force startPC to always load from a snippet
Previously, the assumption was that the value of the startPC symbol
would never be used before being set during binary encoding. However,
since memory instruction expansion is now being moved to happen before
binary encoding, we can no longer rely on the startPC symbol having its
address set prior to expansion of this instruction.

The ideal solution to this would be to use a relocation of some kind to
fill in the immediate values after binary encoding completes. However,
the value of startPC doesn't seem to be used on any performance-critical
code paths, so a temporary solution (albeit with slightly reduced
performance) is to simply always use a constant data snippet.

Signed-off-by: Ben Thomas <ben@benthomas.ca>