Implement expansion of unresolved memory instructions
In order for OpenJ9 to work correctly, it's necessary for a
MemoryReference with an associated UnresolvedSnippet to expand into a
very specific sequence of instructions. For the time being, in order to
avoid breaking OpenJ9, this code is being added to OMR. Eventually, it
should be possible to sink this code into OpenJ9 by making use of the
new expandForUnresolvedSnippet method as an extension point.

Signed-off-by: Ben Thomas <ben@benthomas.ca>