Cache clear and abandon simplification
When creating remainder or abandoning it, use target environment even if
current thread != from target environment.

This we resolve the problem of occasional waste of abandoned Tenure TLH
remainders for mutator threads. Now, we will be able to re-use them on
next CS cycle.

However, to be able to safely use target's environment, it must be
ensured that target thread is not using the remainder at the same time.
This will be ensured by either target thread not holding VM access, or
if racing to acquire VM access it will serialize by external means (for
example, in OpenJ9, by public flags mutex).

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>