Fix zero length array warning on z/OS
The following warning is emitted when compiling a compilation unit
which includes the OMRCodeCache.hpp file:

```
omr/compiler/runtime/OMRCodeCache.hpp", line 450.14: CCN8891 (W) The
array member "void *_CCPreLoadedCode[]" may only be followed by members
of consistent type.
```

The reason this warning is emitted is because `TR_numCCPreLoadedCode`
evaluates to 0 on S390 due to an ifdef and so we have a zero length
array on z/OS. It is emitting a warning because we have types declared
after the zero length array which are not of the same type as the array
elements.

To fix this issue we simply omit an ifdef which forces the value of
`TR_numCCPreLoadedCode` to be non-zero on S390. This is reasonable, as
this enum is not used at all in the S390 codegen.

Fixes: #3703

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>