Reduce scratchSpaceLimit to 16 MB under `-Xtune:virtualized` option
`-Xtune:virtualized` is an option used by a downstream project (openj9)
and geared towards virtualized/cloud environments where low memory and
CPU consumption is a desired feature, even if it comes at the expense of
a small throughput degradation.
During a compilation, the JIT is allowed to allocate memory for its
internal needs up to a `scratchSpaceLimit`, which is currently set at
256 MB. As a side effect, large transient spikes in memory usage can
be triggered, which forces the user to use large (Docker) containers
to account for these possible spikes.
This commit lowers the `scratchSpaceLimit` to 16 MB when
`-Xtune:virtualized` option is in effect. We have measured small
throughput degradations from this change (2-3%) in some Java EE
benchmarks, but also reductions in CPU consumed by compilation threads
in the range of 8-31%. Memory usage spikes due to JIT compilations are
of course capped to 16 MB.
Note that when a compilation fails because it cannot be performed within
the memory space alloted, it will be retried at a lower optimization level
which typically consumes (much) less memory.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>