Do not test shifts wider than the type
Shifting a value by more than bit-width of datatype (or by negative value)
is undefined behavior in C++, therefore we cannot test it using test
oracle (shift_left<>() and shift_right<>()) implemented in C++.

Moreover, semantics of such shifts is not defined for OMR compiler
so testing it makes little sense. This is an ongoing issue (#1498).

As an interrim solution, this commit changes shift tests to only
test shifts within 0..bitwidth-1 range.

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>