Don't overwrite the _callNode of a mapped callsite
findAndUpdateCallSiteInGraph maps the callNode generated in actual ilgen
to a callsite created when doing estimate code size baseed on byte code
info. There is a potential problem if there are more than one callNode at
the same byte code index. For now, we always choose to inline the first
callNode with matching byte code info and argument info.

This change also changes the location of the trace message
"before findAndUpdateCallSiteInGraph" so that it actually appears
before the callsite gets updated.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>