Fix marking of failed tests in Azure Pipelines
When a test fails we want the job to be marked as failed as well,
however we still wish to run the "Publish results" job so that the test
output can be parsed and test results published. We modify the pipeline
so if there are failures we fail the "Test" job, but still continue to
run the "Publish results" job.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>