ClearCache on deactivate
For Concurrent Scavenger, when a Mutator releases VM access, we now
clear the cache eagerly, even though it might just be deactivated and
not pushed to scanning yet.

Otherwise, deferring the clearing at the point of pushing an inactive
cache may cause a subtle race with a third thread doing it on behave of
owner thread, later toward the end of concurrent cycle.

With this change clearCache can only occur in the context for owning
thread, and never in the context of a third thread (signaling GC thread
for final scanning, or Mutator thread doing exclusive VM access thread
pass)!

Small downside is that we may clear the cache (effectively check and
create if there is a potential remainder) multiple times during CS
cycle, before it's finally abandoned.

Newly introduced INACTIVE cache flag is just for debugging purposes.

Some other smaller changes piggybacked:
- removed unused canLocalCacheBeReused
- removed stale/incorrect assert in isConcurrentWorkAvailable()
- wait in shouldDoFinalNotify should be 1ms, not 1000ms

More specifics of the race being resolved (with some details from
OpenJ9):
One mutator thread (M1) doing flush on another mutator thread (M2)
(through signalThreadsToFlushCaches toward the end of concurrent phase).
M1 won race to flush inactive Survivor/Tenure cache, and as a part of it
(before pushing cache for scanning) it's creating remainders, what is
done in 2 steps: set Alloc first and Top second, non atomically.
M2 used to be blocked (no VM access) and just woke up (not necessarily
through fast path, hence acquiring public flag mutex and not being able
to synchronize with M1) to execute a RB and wants to copy an object. It
lost race to activate inactive cache, however it sees there is a
remainder, but since it's racing with M1 it sees it only half
initialized (Alloc set, but Top not) and asserts.


Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>