Implement AOT relocation during MemoryReference expansion
The previous implementation of binary encoding for instructions with
MemoryReferences would sometimes add AOT relocations. It's not fully
known how much of this code was in use, but for the sake of consistency
for the time being, the new implementation will now also add AOT
relocations during MemoryReference expansion if necessary.

Signed-off-by: Ben Thomas <ben@benthomas.ca>