Re-initialize method builder objects before inlining
When using the JitBuilder Call service and passing a MethodBuilder
object, we do not know the origin of that object before we begin
to inline it. In most code samples, that object is a "fresh"
MethodBuilder that hasn't been compiled itself. When it has been
compiled previously, there are some TR objects cached in it that
cause problems during inlining, such as the compilation object,
the method symbol, etc. These should be re-initialized to the
current MethodBuilder being compiled before the callee MethodBuilder
is inlined. While setupForBuildIL() is called, we also need to
call initialize() to re-initialize these fields.

Included in this fix is an update to the Call code sample from
@fjeremic that causes that sample to test this exact scenario
(inlining a MethodBuilder object that has previously been
compiled).

Fixes #4909

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>