Fix max open file limits on OSX
on OSX, there are two mechanisms for reporting the limits of the system.
The first is getrlimit/setrlimit, and the second is through sysctl.

Sysctl will report the maximum number of open files that the kernel will
allow a process to have. This is a global property that affects all
processes.

getrlimit/setrlimit, on the other hand, will report the limits set on a
given process.

These two limits do not have to match.

The true soft limit for open files is the minimum between the current
rlimit and the sysctl limit.  The true hard limit is the minimum between
the max rlimit and the sysctl limit.

With this PR, this is what we report and enforce on OSX.

Signed-off-by: Robert Young <rwy0717@gmail.com>