Add support to indicate synthetic exception handler
Sometimes compiler can choose to insert exception handlers that do not
exit in the source code. Special handling of those handler
might be necessary. In case of OSR, the runtime would need to
handle synthetic handler differently than a normal handler.

This PR adds support to indicate certain block or handler is for
synthetic exception handler.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>