Write out newline character after the message in verbose output
The `TR_VerboseLog::writeLine*` APIs currently write the newline
character before the message being printed which is non-conventional.
Typically the expectation of a user is that any API that outputs
newline characters outputs them at the end, examples from various
languages in [1] [2] [3], and there are numerous other examples.

The reason we are making this change is because other components in OMR
will expect to be writing at the start of a newline, whether that be
stdout, stderr, or a file. The verbose log not adhering to this
expectation yields in messages getting printed one after another which
is not desirable.

[1] https://docs.python.org/3/library/functions.html#print
[2] https://docs.oracle.com/javase/8/docs/api/java/io/PrintStream.html#println-java.lang.String-
[3] https://docs.microsoft.com/en-us/dotnet/api/system.console.writeline?view=netframework-4.8#System_Console_WriteLine_System_String_

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>