AArch64: Add check before using register of node in memory reference
`OMR::ARM64::MemoryReference::addToOffset` uses node's register as displacement
if node is `const` node. It uses the node associated with the last instruction
if `NULL` is passed to `addToOffset` as `node` argument. Unfortunately,
the node chosen is irrelevant in some cases.
This commit adds value check before using the node's register to avoid using
incorrect register for displacement.
This commit also moves if-block for the case where const value can be encoded
in `addimmx` instruction before the if-block for the case where node's register
can be reused.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>