Prevent TR_X86ProcessorInfo::initialize from re-initializing
The static variable TR_X86ProcessorInfo should only be initialized
once. Initializing it repeatly caused a race condition in the code.
Technically we may still have syncronization issues without a lock,
but the chances will be much lower with this change. Note that this
implementation of processor detection will be replaced soon. This change is only needed to test out the new processor detection APIs.

Issue: #4339

Signed-off-by: Harry Yu <harryyu1994@gmail.com>