Fix incorrect spelling of J9_PROJECT_SPECIFIC
It seems that previously, there was code in the Power codegen guarded by
the definition of a J9_ProjectSpecific macro. This was a misspelling of
the J9_PROJECT_SPECIFIC macro, so this code was never actually used.
This could cause incorrect behaviour in OpenJ9 if an irsload node
appeared under an s2i node. The spelling of this check has been
corrected so that the code should now be enabled correctly.

Fixed: #5137
Signed-off-by: Ben Thomas <ben@benthomas.ca>