Deprecate TR_TraceCGEvaluation
This tracing is not enabled via `TR_TraceAll` or via `TR_TraceCG`.
Looking at the output of enabling this tracing I do not see the value
of it as all of the information can be extracted from what is printed
during instruction selection. For example, the live registers are
printed before and after evaluation of a treetop so we can extract
which registers got killed from there, and node evaluation tracing does
not seem to provide anything useful really as it prints killed
registers primarily and the tracing is quite verbose when we come
across dummy registers.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>