Deprecate TR_TraceVIP with TR_TraceCG
The `TR_TraceVIP` is unused in the inliner so it can be completely
removed there. The only other place it is used is in the Z codegen when
tracing BCD evaluation. OMR currently does not have BCD trees, and
these are only exploited in the OpenJ9 project. As such the code should
move there, but irregardless of that fact the tracing is really
non-invasive and only related to specific BCD optimization which only
gets triggered in a small set of methods in the Data Access Accelerator
library. Enabling the tracing by `TR_TraceCG` is acceptable here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>