Deprecate TR_TraceRegisterState
This option is currently only available in the Z codegen. It calculates
a bitvector of used vs. free registers and it adds it as an instruction
comment. I do not recall ever using this option and cannot envision a
scenario in which it would be more useful than `TR_TraceRA`. The
operation is quite expensive as well requiring an O(N) pass for each
instruction and dynamic memory allocation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>