Deprecate code mining related tracing options
Code mining is a Z codegen only concept in which annotations are added
as comments to certain objects such as instructions and blocks for post
processing in a tool called the "Visual Performance Analyzer". Code
mining has not been used in almost a decade however and we do not
envision making use of the tool in the future either. As such all the
related code is slated for deprecation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>