Fixed imm value for addis and lis
addis and lis accept a signed 16 bit value as an immediate. They are passed
into generateTrg1Src1ImmInstruction as signed 32 bit values. In some edge
cases, the sign extension was not handled properly. In most cases this was due
to the value being incremented after sign extension. In other cases it was due
to an unnecessary mask clearing the upper 16 bits.

Also fixed an issue in inlineSimpleAtomicUpdate where the check for whether a
value stored as a signed 64 bit could fit in a signed 32 bit was incorrect.

Closes: #5212
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>