Deprecate TR_TraceUnsafeInlining with debug level trace inlining
The unsafe inlining tracing currently uses a custom tracing option
needlessly. The tracing provided seems a bit verbose, but we do not see
a reason why we should complicate things by using a separate tracing
option. Instead we fold unsafe inlining tracing into the debug level
tracing which is already available.

The tracing guarded by the option being deprecated seems to be
incomplete. There are many trace statements which have "UsafeInlining:"
in them, however they are not guarded by the above tracing option. We
get rid of this inconsistency here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>