Deprecate TR_TraceTempUsage with envvar
This tracing option counts the number of stores to temp slots by doing
a linear walk of the treetops. It seems overkill to have two separate
tracing optimizations for this, and it seems like something that would
be used during debugging anyways as this information can be obtained
otherwise via a simple grep.

In light of this we deprecate the two options related to tracing temp
counts, and rather than deleting the code we leave it under an envvar.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>