Skip all tests for NaN values on zOS
On zOS, NaN values are formatted as `NaNQ(1)` when given to
std::snprintf(). This makes it difficult to parse in TRIL parser.
Instead of adding yet another workaround, this commit skips
all tests involving NaN constants on zOS until a better solution
to create TRIL code is implemented (see issue #5183)

Signed-off-by: Jan Vrany <jan.vrany@fit.cvut.cz>