Make UnsupportedValueTypeOperation a recoverable exception
The UnsupportedValueTypeOperation exception extended
CompilationException, which meant that, if it was thrown, compilation
would be aborted.  However, it should be considered a recoverable
exception - if an unsupported operation is encountered during
inlining, only the inlining of that particular method should be aborted
while compilation of the outermost method continues.

Fixed this by having UnsupportedValueTypeOperation extend
RecoverableILGenException instead.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>