Consider value type arrays in VPHandler for ArrayStoreCHK
In prototype support for value types, the AASTORE bytecode will throw a
NullPointerException if the source operand is a null reference.  This
checking will be performed under ArrayStoreCHK.  The VPHandler for
ArrayStoreCHK needs to consider whether the source of the store could be
a null reference and the component type of the array could be a value
type - if so, the ArrayStoreCHK cannot be removed, so the necessary
checking can be performed.

In addition, NullPointerException needs to be included in the table of
_standardExceptions thrown by ArrayStoreCHK.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>