Create an extensible Peephole phase
This phase provides the necessary plumbing for codegens to extend to
perform peephole optimizations. Currently it is quite primitive in that
we do a linear walk over the instructions and we let the codegens
implement the specific peepholes for the target architecture. However
in the future we envision improving the common peephole plumbing to
support peephole windows, etc. which all the codegens will
automatically benefit from using a shared codebase.

In this commit we setup the plumbing to migrate the Z peephole
optimizations over to the new Z extension. In following commits we will
perform the same exercise on other platforms.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>