Fixes in RegDepCopyRemoval for postGRA block splitter
RegDepCopyRemoval is the last ran optimization which intends to reduce
register shuffling due to global register dependencies. It was
generating tree where we have a PassThrough Child under PassThrough to
copy node to new virtual register. This tree was not working well with
the post GRA block splitter. Issue was that PassThrough child
encountered in GlRegDeps was anchored to tree top. In postGRA block
splitter, it records the regStores encountered before the GlRegDeps in
extended basic block to guide decision on which register to chose for
the node after split point.In this commit, adding changes to both
splitPostGRA block splitter and RegDepCopyRemoval optimization so that
we can split the blocks after later has run.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>