Deprecate DAA related peepholes
The DAA related peepholes can be summarized as follows:

- Removal of redundant `BRC` instructions
- Removal of redundant spill
- Inlining of `EX` by creating unconditional branch

The first two no longer happen in DAA as the upstream code (which
actually lives in the OpenJ9 project) has been fixed to no longer
generate these redundant branches.

The last optimization is dubious at best. It attempts to look for an
unconditional branch following which it can inline the `EX` target so
we can change the `EX` to an `EXRL`. Not only is this highly
improbable, but the fallback the peephole does is to insert an
unconditional branch itself which doesn't seem forward thinking. I
checked the codebase and we don't really generate `EX` dispatches
anymore due to availability of `EXRL`, so rather than maintain this
rather dubious peephole optimization.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>