Place exceptions for value types under object model check
Rather than using an ifdef of whether value types are enabled in
OMR::Block::_standardExceptions to select the set of checks that can
throw a NullPointerException, instead added any additional checks to
a _valueTypesExceptions array.  These are added to the set of checks
whose exceptions are caught in OMR::Block::setExceptionClassName if
value types are enabled in the object model.  Added a corresponding
object model method, areValueTypesEnabled, whose base implementation
returns false.

Similarly, added CanCatchNullCheck in OMR::Node::exceptionsRaised to
the set of possible exceptions thrown by ArrayStoreCHK if value types
are enabled in the object model.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>