Deprecate S390Peephole in favor of the new extensible Peephole phase
The rest of the code in S390Peephole is now deprecated and not needed.
The only thing that was left in there was a z10 true complement
peephole which tried to address a hardware performance issue. The
shortcoming of that peephole is that it didn't actually do anything
because it attempts to insert NOP `LR GPRX,GPRX` instructions into the
instruction stream, however subsequent peephole optimizations remove
such NOPs so the effect is nil.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>