Set previous and next pointers to NULL when removing an instruction
The peephole codegen phase relies on an assumption that if an operation
was performed on an instruction, and that instruction got removed from
the instruction stream, then the instruction has been completely
unlinked from the instruction linked list.

This is in fact true today, however the removal from the linked list
only happens one-way, i.e. the removed instruction is not reachable
from the instruction linked list, however the removed instruction
itself still holds pointers back to the location where it was
originally. This is problematic as functions may have locally cached
the pointer to the removed instruction and cannot any longer determine
whether that pointer is in the instruction stream or not.

To remedy this we simply do a two-way removal by updating the previous
and next pointers to be `NULL`.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>