Remove some statics from z/codegen
Downstream projects such an OpenJ9's JITServer
cannot operate correctly if certain fields are static.
This is because the JITServer can talk to multiple
JVM clients such that the static field's value on the JITServer
side is no longer valid for all client JVMs. This commit
removes those problematic statics.

An unused function is also removed.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>