Add prefix field to instruction metadata
Previously, the template instruction for each instruction was stored in
the opcode field of the metadata table. However, POWER10 introduced the
concept of prefixed instructions, which require a 64-bit template
instruction. In order to accomodate this, a new prefix field has been
added to the instruction metadata table to hold the prefix word of the
template instruction.

Signed-off-by: Ben Thomas <ben@benthomas.ca>