Change XPLINK call descriptor compilation failure to fatal assert
Since the change in #4890 we now always encode the XPLINK call
descriptor inline right after the call to handle large methods which
are common on z/OS 31-bit. To avoid silent performance problems of
outside factors we turn the compilation failure into a fatal assert to
ensure we never hit this code path.

The reason we are making this change here is due to the previous
commit. We had a bug where we remove an instruction (NOP alignment)
while binary encoding the instruction itself. Since we now nullify the
previous and next pointers the binary encoder will stop encoding
subsequent instructions. It then proceeds to try and encode the
snippets and it fails because the pointers are not properly setup.
These were silent failures as the compilation would fail, however we
are still able to proceed.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>